pyproject.toml
[project]
name = "flight-display"
version = "0.1.0"
description = "FlightRadar24 display for Raspberry Pi"
authors = [
    {name = "Rory Graves"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[project.scripts]
flight-display = "flight_display.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
